requirements.txt
streamlit
langchain_openai
python-dotenv
langchain_community 
langchain_core
langserve
fastapi
uvicorn
sse_starlette